Отделена реализация сервисов от контракта
